+++ cli/app_metadata.py
@@ -1,5 +1,5 @@
 
 class Metadata:
-    __version__ = "2.0.3"
+    __version__ = "2.0.4"
     __author__ = "Zap-09"