+++ inkpull/scraper/atsumaru/parsing.py
@@ -1,5 +1,7 @@
+import sys
 from utils import log
 from .exceptions import AtsumaruException
+from .schemas import AtsumaruScanlator, AtsumaruChapter
 
 
 def clean_url(url: str) -> str:
@@ -151,21 +153,194 @@ def get_manga_id(text: str) -> str:
         raise AtsumaruException.MangaIDNotFound(text)
 
 
-def make_chapter_urls(manga_id: str, text: dict, scanlation_id: str | None) -> list:
-    chapters = text.get("chapters", None)
+def construct_chapter_urls(manga_id: str, data: list[AtsumaruChapter]) -> list[AtsumaruChapter]:
+    comic_to_download = []
 
-    if chapters is None:
-        raise AtsumaruException.UnexpectedJsonStructure("chapters")
+    for chapter in data:
+        chapter.chapterUrl = f"https://atsu.moe/read/{manga_id}/{chapter.id}"
+        comic_to_download.append(chapter)
+    return comic_to_download
 
-    chapter_urls = []
-    for chapter in chapters:
-        chapter_scanlation = chapter.get("scanlationMangaId")
 
-        if scanlation_id is None or chapter_scanlation == scanlation_id:
-            chapter_urls.append(
-                f"https://atsu.moe/read/{manga_id}/{chapter["id"]}"
+def parse_user_input(user_input: str, max_items: int) -> list[int]:
+    args = user_input.strip()
+
+    if not args:
+        raise ValueError(
+            "Invalid input. Empty input."
+        )
+
+    if args == "--help":
+        log("Use numbers to pick scan groups.", "info")
+        log("Example: 1,3", "info")
+        log("Use --all to select all.", "info")
+        log("Use 0 to exit.", "info")
+        raise AtsumaruException.HelpFlag
+
+    if args == "--all":
+        return list(range(max_items))
+
+    if args == "0":
+        raise SystemExit
+
+    selection_seen = set()
+    selection = []
+
+    for item in args.split(","):
+        item = item.strip()
+
+        if not item.isdigit():
+            log(f"{item} is not a number, skipping...", "warn")
+            continue
+
+        number = int(item)
+
+        if number == 0:
+            raise ValueError(
+                "0 cannot be combined with other selections"
             )
-    if len(chapter_urls) <= 0:
-        raise AtsumaruException.NoChaptersToDownload("Chapter_url list is less than 1")
 
-    return chapter_urls
+        index = number - 1
+
+        if index < 0 or index >= max_items:
+            raise IndexError(
+                f"{number} is out of range"
+            )
+
+        if index not in selection_seen:
+            selection_seen.add(index)
+            selection.append(index)
+
+    if not selection:
+        raise ValueError(
+            "No valid selections"
+        )
+
+    return selection
+
+
+def extract_scanlators(all_chapter_json: dict, page_json: dict) -> list[AtsumaruScanlator]:
+    chapters_object = all_chapter_json.get("chapters", None)
+    if not chapters_object:
+        raise AtsumaruException.UnexpectedJsonStructure(
+            f"Unexpected json structure. Can't find key 'chapters'"
+        )
+    chapters_object.reverse()
+    chapters = [
+        AtsumaruChapter(
+            id=c["id"],
+            scanlationMangaId=c["scanlationMangaId"],
+            title=c["title"],
+            number=c["number"],
+        )
+        for c in chapters_object
+    ]
+
+    chapter_by_scanlator = {}
+
+    for c in chapters:
+        chapter_by_scanlator.setdefault(
+            c.scanlationMangaId, []
+        ).append(c)
+
+    manga_page = page_json.get("mangaPage", None)
+
+    if manga_page is None:
+        raise AtsumaruException.UnexpectedJsonStructure(
+            f"Unexpected json structure. Can't find key 'mangaPage'"
+        )
+    scanlators_object = manga_page.get("scanlators", None)
+    if not scanlators_object:
+        raise AtsumaruException.UnexpectedJsonStructure(
+            f"Unexpected json structure. Can't find key 'scanlators'"
+        )
+
+    scanlators = [
+        AtsumaruScanlator(
+            id=s["id"],
+            name=s["name"],
+            chapters=chapter_by_scanlator.get(s["id"], [])
+        )
+        for s in scanlators_object
+    ]
+    return scanlators
+
+
+def smart_select_chapters(total_items: list[AtsumaruScanlator]) -> list[AtsumaruChapter]:
+    seen_chapters = set()
+    selected_chapters = []
+
+    for group in total_items:
+        for chapter in group.chapters:
+            if chapter.number not in seen_chapters:
+                chapter.scanlationGroupName = group.name
+                seen_chapters.add(chapter.number)
+                selected_chapters.append(chapter)
+            else:
+                continue
+
+    return selected_chapters
+
+
+def flatten_scanlator_chapters(total_items: list[AtsumaruScanlator]) -> list[AtsumaruChapter]:
+    selected_chapters = []
+    for group in total_items:
+        for chapter in group.chapters:
+            chapter.scanlationGroupName = group.name
+            selected_chapters.append(chapter)
+
+    return selected_chapters
+
+
+def interactive_selection(scanlators: list[AtsumaruScanlator]
+                          ) -> list[AtsumaruChapter]:
+    log("Use comma ',' if you want to pick more than one scan group. Eg; 1, 4")
+    while True:
+        log("0. Exit", "info")
+
+        for idx, scanlator in enumerate(scanlators, start=1):
+            log(
+                f"{idx}. "
+                f"{scanlator.name} "
+                f"({scanlator.chapter_count} chapters)", "info"
+            )
+
+        user_input = input(
+            "Select scanlators: "
+        )
+        try:
+            indexes = parse_user_input(
+                user_input,
+                len(scanlators)
+            )
+
+            selected_scanlators = [
+                scanlators[i]
+                for i in indexes
+            ]
+            return flatten_scanlator_chapters(selected_scanlators)
+
+        except AtsumaruException.HelpFlag:
+            pass
+        except SystemExit:
+            log("Exiting...", "info")
+            sys.exit(0)
+        except IndexError as e:
+            log(str(e), "error")
+        except ValueError as e:
+            log(str(e), "error")
+
+
+def resolve_selection(scanlators: list[AtsumaruScanlator],
+                      *,
+                      select_all: bool = False,
+                      smart_select: bool = False,
+                      ) -> list[AtsumaruChapter]:
+    if smart_select and select_all:
+        log("Both --all and --smart flag are present. Ignoring --smart", "warn")
+
+    if select_all:
+        return flatten_scanlator_chapters(scanlators)
+    elif smart_select:
+        return smart_select_chapters(scanlators)
+    return interactive_selection(scanlators)